refactor: store/index.ts 파일 분리 - index.ts는 export만 선언 - headerStore.ts에 기능 이전
ref: #5
