+++ src/shared/store/index.ts
@@ -1,12 +1,2 @@
 // index.ts
-import { create } from 'zustand';
-
-interface HeaderState {
-  headerHeight: number;
-  setHeaderHeight: (height: number) => void;
-}
-
-export const useHeaderStore = create<HeaderState>((set) => ({
-  headerHeight: 0,
-  setHeaderHeight: (height) => set({ headerHeight: height }),
-}));
+export { useHeaderStore } from './headerStore';
